app/requirements.txt
appnope==0.1.2
asgiref==3.3.4
backcall==0.2.0
certifi==2021.5.30
decorator==5.0.9
Django==3.0.1
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.1
django-environ==0.4.5
django-filter==2.4.0
django-nine==0.2.4
djangorestframework==3.12.4
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
gunicorn==20.0.4
ipdb==0.13.9
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Markdown==3.3.4
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2-binary==2.8.5
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
